% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opdyke.percentiles.R
\name{opdyke.percentiles}
\alias{opdyke.percentiles}
\title{Calculate the Opdyke percentile of each predicted correlation from a structural equation model}
\usage{
opdyke.percentiles(fit = NULL, lower = 0.4, upper = 0.6, precision = "less")
}
\arguments{
\item{fit}{a lavaan object containing results from a fitted structural equation model. This will contain the observed and predicted covariance
and correlation matrices needed to calculate Opdyke percentiles.}

\item{lower}{The minimum percentile of the Opdyke distribution that is considered to be acceptably close to the observed correlation (represented by the Opdyke
distribution median). The percentile of the predicted correlation will be compared to this value in the simplified output to determine if the model reasonably
reproduces an observed correlation. The default is 0.40.}

\item{upper}{the maximum percentile of the Opdyke distribution that is considered to be acceptably close to the observed correlation (represented by the Opdyke
distribution median). The percentile of the predicted correlation will be compared to this value in the simplified output to determine if the model reasonably
reproduces an observed correlation. The default is 0.60.}

\item{precision}{Controls the precision of the probability density function and cumulative distribution function calculations. The default is precision= “less”
which calculates the PDF and CDF for polar angles between (0,pi) in .01 increments. precision= “more” calculates the PDF and CDF for polar angles between (0,pi)
in .001 increments, which takes considerably longer, especially if there are many correlation elements.}
}
\value{
Opdyke percentiles of predicted correlations from a structural equation model
}
\description{
This function aids is determining the discrepancy between an observed correlation and a predicted correlation from a structural equation model.
The function calculates the percentile of the predicted correlation within the observed correlation's Opdyke distribution. The output shows elements with
large discrepancies, the percentiles associated with all elements, the bound of each element for which the correlation matrix remains positive definite, and the
interval of correlation residuals that are between the range specified by the "upper" and "lower" arguments.
}
\examples{
#assign dataset to “dat”
dat<-lavaan::HolzingerSwineford1939
#define a two-factor model with correlated factors
model<-"f1=~ x1+x2+x3
f2 =~ x4+x5+x6
f1~~f2"
#fit the two-factor model in lavaan
f<-lavaan::cfa(data=dat, model=model)
#calculate Opdyke percentiles for each predicted correlation element
OP<-opdyke::opdyke.percentiles(fit=f, lower=.40, upper=.60)
}
\author{
Daniel McNeish

Maintainer: Daniel McNeish \href{mailto:dmcneish@asu.edu}{dmcneish@asu.edu}
}
