% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opdyke.R
\name{opdyke}
\alias{opdyke}
\title{Calculate the Opdyke distribution PDF and CDF  for each element in a correlation matrix.}
\usage{
opdyke(cor = NULL, type = "lavaan", precision = "less")
}
\arguments{
\item{cor}{An object containing the observed correlation matrix. Either a raw correlation or a lavaan object. Object type is specified in the 'type' option.
this is a required argument for the function.}

\item{type}{Specifies the type of object in the cor argument. The default is type='lavaan'. The other acceptable option is type='cor' if the object is a
raw correlation matrix.}

\item{precision}{Controls the precision of the probability density function and cumulative distribution function calculations. The default is precision= “less”
which calculates the PDF and CDF for polar angles between (0,pi) in .01 increments. precision= “more” calculates the PDF and CDF for polar angles between (0,pi)
in .001 increments, which takes considerably longer, especially if there are many correlation elements.}
}
\value{
list of Opdyke distribution PDF and CDF values for each element of a correlation matrix
}
\description{
This function calculates the probability density function and cumulative distribution function of the Opdyke distribution for
value between 0 and pi. The output is a list where each list element is a different correlation matrix from the data. This function is useful for creating
custom plots or custom calculations that involves the Opdyke distribution.
}
\examples{
#assign dataset to “dat”
dat<-lavaan::HolzingerSwineford1939
#assign correlation matrix to “A”
R<-cor(dat[,7:12])
#create lower triangle matrix of polar angles
D<-opdyke::opdyke(cor=R, type="cor")
#print rows 100 to 125 of the 10th list element
D[[10]][100:125,]
}
\author{
Daniel McNeish

Maintainer: Daniel McNeish \href{mailto:dmcneish@asu.edu}{dmcneish@asu.edu}
}
