% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2polar.R
\name{r2polar}
\alias{r2polar}
\title{Convert correlation matrix to polar angles}
\usage{
r2polar(cor = NULL, type = "lavaan")
}
\arguments{
\item{cor}{An object containing the observed correlation matrix. Either a raw correlation or a lavaan object. Object type is specified in the 'type' option.
this is a required argument for the function.}

\item{type}{Specifies the type of object in the cor argument. The default is type='lavaan'. The other acceptable option is type='cor' if the object is a
raw correlation matrix.}
}
\value{
Lower triangle of polar angles.
}
\description{
This function transforms a symmetric correlation matrix to a lower triangle of polar angles
Input can be a raw correlation matrix or a lavaan object.
}
\examples{
#assign dataset to “dat”
dat<-lavaan::HolzingerSwineford1939
#assign correlation matrix to “A”
R<-cor(dat[,7:12])
#apply r2polar function with type= “cor”
r2polar(cor=R, type="cor")
}
\author{
Daniel McNeish

Maintainer: Daniel McNeish \href{mailto:dmcneish@asu.edu}{dmcneish@asu.edu}
}
